+++ src/main/java/hr/dlatecki/algorithms/gen_alg/codecs/interfaces/package-info.java
@@ -4,4 +4,4 @@
  * @author Domagoj Latečki
  * @since 1.8
  */
-package hr.dlatecki.algorithms.gen_alg.codecs;
+package hr.dlatecki.algorithms.gen_alg.codecs.interfaces;