+++ src/main/java/hr/dlatecki/algorithms/gen_alg/impl/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Package which contains implementations of the genetic algorithms.
- * 
- * @author Domagoj Latečki
- * @since 1.8
- */
-package hr.dlatecki.algorithms.gen_alg.impl;