+++ src/main/java/hr/dlatecki/algorithms/gen_alg/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Package which contains implementations of the genetic algorithms. It also contains sub-packages which define
+ * population and operators in the algorithms.
+ * 
+ * @author Domagoj Latečki
+ * @since 1.8
+ */
+package hr.dlatecki.algorithms.gen_alg;