+++ src/main/java/hr/dlatecki/algorithms/gen_alg/population/chromosomes/DoubleArrayChromosome.java
@@ -1,4 +1,4 @@
-package hr.dlatecki.algorithms.gen_alg.population;
+package hr.dlatecki.algorithms.gen_alg.population.chromosomes;
 
 import java.util.Arrays;
 import hr.dlatecki.algorithms.gen_alg.population.abstracts.AbstractChromosome;