+++ src/main/java/hr/dlatecki/algorithms/gen_alg/population/chromosomes/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Package which contains implementations of the chromosomes used in genetic algorithms defined in
+ * {@link hr.dlatecki.algorithms.gen_alg} package.
+ * 
+ * @author Domagoj Latečki
+ * @since 1.8
+ */
+package hr.dlatecki.algorithms.gen_alg.population.chromosomes;